Put IoT tests in separate file
