+++ test/kitchen/test/integration/dd-agent-install-script/rspec/dd-agent-install-script_spec.rb
@@ -1,16 +1,5 @@
 require 'spec_helper'
-
-# We retrieve the value defined in kitchen.yml because there is no simple way
-# to set env variables on the target machine or via parameters in Kitchen/Busser
-# See https://github.com/test-kitchen/test-kitchen/issues/662 for reference
-let(:agent_flavor) {
-  if os == :windows
-    dna_json_path = "#{ENV['USERPROFILE']}\\AppData\\Local\\Temp\\kitchen\\dna.json"
-  else
-    dna_json_path = "/tmp/kitchen/dna.json"
-  end
-  JSON.parse(IO.read(dna_json_path)).fetch('dd-agent-rspec').fetch('agent_flavor')
-}
+require 'iot_spec_helper'
 
 shared_examples_for 'Agent installed by the install script' do
   context 'when testing DD_SITE' do
@@ -42,13 +31,13 @@ shared_examples_for 'Agent installed by the install script' do
 end
 
 describe 'dd-agent-installation-script' do
-  include_examples 'Agent install'
-
   if agent_flavor == "datadog-agent"
+    include_examples 'Agent install'
     include_examples 'Agent behavior'
     include_examples 'Agent installed by the install script'
     include_examples 'Agent uninstall'
   elsif agent_flavor == "datadog-iot-agent"
+    include_examples 'IoT Agent install'
     include_examples 'IoT Agent behavior'
     include_examples 'Agent installed by the install script'
     include_examples 'IoT Agent uninstall'
