+++ test/kitchen/test/integration/dd-agent-install-script/rspec/iot_spec_helper.rb
@@ -1,0 +1,1 @@
+../../common/rspec/iot_spec_helper.rb
